Rework test directory setup/teardown.

+++ tests/unit/test_worker.py
@@ -172,40 +172,39 @@ class TestWorkerClass():
         srcDir = "./tests/data/in/testDir"
         destDir = "./tests/data/out/testDir"
         # make sure outputDir does not exist
-        try:
+        if os.path.exists(destDir):
             os.rmdir(destDir)
-        except Exception:
-            pass
-        # create testDir
-        try:
+        # create srcDir
+        if os.path.exists(srcDir) is False:
             os.mkdir(srcDir)
-        except Exception:
-            pass
         assert os.path.isdir(srcDir) is True
         assert os.path.isdir(destDir) is False
         worker = Worker()
         worker.rename_directory(srcDir, destDir)
         assert os.path.isdir(srcDir) is False
         assert os.path.isdir(destDir) is True
-        # cleanup test and output dirs
-        try:
+        # cleanup input and output dirs
+        if os.path.exists(srcDir):
             os.rmdir(srcDir)
-        except Exception:
-            pass
-        try:
+        if os.path.exists(destDir):
             os.rmdir(destDir)
-        except Exception:
-            pass
 
     # test rename_directory function failure.
     # failure occurs when src_dir does not exist
     def test_rename_directory_failure_src_not_exists(self):
         srcDir = "./tests/data/in/random_dir_name"
         destDir = "./tests/data/out/testDir"
+        if os.path.exists(srcDir):
+            os.rmdir(srcDir)
+        if os.path.exists(destDir):
+            os.rmdir(destDir)
         assert os.path.exists(srcDir) is False
+        assert os.path.exists(destDir) is False
         worker = Worker()
         try:
             worker.rename_directory(srcDir, destDir)
+            # exception should have been thrown, but wasn't
+            assert False
         except Exception as e:
             assert "No such file or directory" in str(e)
 
@@ -214,14 +213,12 @@ class TestWorkerClass():
     def test_rename_directory_failure_dest_exists(self):
         srcDir = "./tests/data/in/testDir"
         destDir = "./tests/data/out/testDir"
-        assert os.path.exists(srcDir) is False
-        assert os.path.exists(destDir) is False
-        # create testDir
-        try:
-            os.mkdir(srcDir)
-        except Exception:
-            pass
-        os.makedirs(destDir)
+        # create srcDir
+        if os.path.exists(srcDir) is False:
+            os.makedirs(srcDir)
+        # create destDir
+        if os.path.exists(destDir) is False:
+            os.makedirs(destDir)
         assert os.path.exists(srcDir) is True
         assert os.path.exists(destDir) is True
         worker = Worker()
@@ -229,13 +226,8 @@ class TestWorkerClass():
             worker.rename_directory(srcDir, destDir)
         except Exception as e:
             assert "Destination directory " + destDir + " exists" in str(e)
-            print(e)
-        # cleanup test and output dirs
-        try:
+        # cleanup input and output dirs
+        if os.path.exists(srcDir):
             os.rmdir(srcDir)
-        except Exception:
-            pass
-        try:
+        if os.path.exists(destDir):
             os.rmdir(destDir)
-        except Exception:
-            pass
